updated to work with latest java-jwt 2.1.0 release

+++ src/main/java/com/auth0/spring/security/auth0/Auth0TokenHelper.java
@@ -2,7 +2,7 @@ package com.auth0.spring.security.auth0;
 
 public interface Auth0TokenHelper<T> {
 
-	public String generateToken(T object, int expiration); 
+	public String generateToken(T object, long expiration); 
 	public T decodeToken(String token);
 
 }
